reorganizing the folders
