--- PDFExtract/.classpath
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-        <classpathentry kind="src" path="appClientModule"/>
-        <classpathentry kind="src" path="resources"/>
-        <classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
-                <attributes>
-                        <attribute name="owner.project.facets" value="java"/>
-                </attributes>
-        </classpathentry>
-        <classpathentry kind="con" path="org.eclipse.jst.j2ee.internal.module.container"/>
-        <classpathentry kind="lib" path="lib/pdfbox-2.0.15.jar"/>
-        <classpathentry kind="lib" path="lib/pdf2dom-1.7.jar"/>
-        <classpathentry kind="lib" path="lib/annotations-2.0.1.jar"/>
-        <classpathentry kind="lib" path="lib/commons-io-2.4.jar"/>
-        <classpathentry kind="lib" path="lib/commons-lang3-3.4.jar"/>
-        <classpathentry kind="lib" path="lib/commons-logging-1.2.jar"/>
-        <classpathentry kind="lib" path="lib/fontbox-2.0.0.jar"/>
-        <classpathentry kind="lib" path="lib/FontVerter-1.2.19.jar"/>
-        <classpathentry kind="lib" path="lib/guava-15.0.jar"/>
-        <classpathentry kind="lib" path="lib/javassist-3.18.2-GA.jar"/>
-        <classpathentry kind="lib" path="lib/reflections-0.9.9.jar"/>
-        <classpathentry kind="lib" path="lib/ch.qos.logback.classic-0.9.28.jar"/>
-        <classpathentry kind="lib" path="lib/ch.qos.logback.core-0.9.28.jar"/>
-        <classpathentry kind="lib" path="lib/slf4j.jar"/>
-        <classpathentry kind="output" path="build/classes"/>
-</classpath>
