--- Test/PDFExtract.json
@@ -1,61 +1,0 @@
-{
-"script" : {
-	"sentence_join" : "/var/www/html/experiment/sentence-join/git/sentence-join.py"
-}, 
-"language" : 
-[
-	{
-		"name" : "common",
-		"config" : {
-			"join_words" : [
-				[".*[\\,\\&\\;\\:]$", "", " "],
-				[".*[a-z]+\\-$", "^[a-z]+.*", ""],
-				[".*[a-z]{1,}$", "^[a-z]+.*", " "],
-				[".*[\\,\\;\\s][A-Z]{1,1}$", "", " "],
-				[".*\\s(to|for|at|by)$", "", " "]
-			],
-			"absolute_eof" : [
-				[".*(?<!\\,|\\&|\\;|\\:|\\s[A-Z]{1,1})$", "^[0-9 ]{0,}[A-Z]+.*"],
-				[".*(\\?\\\"?|\\!\\\"?)$", ""],
-				[".*\\w\\.$", ""],
-				["", "^[•]+.*"]
-			],
-			"normalize" : [
-				[ "ﬀ", "ff"],
-				["ﬁ\\s?", "fi"],
-				["ﬂ\\s?", "fl"],
-				["ﬃ", "ffi"],
-				["ﬄ", "ffl"],
-				["ﬅ", "ft"],
-				["ﬆ", "st"],
-				["[“”]", "\""],
-				["[’´]", "'"],
-				["…", "..."],
-				["–", "-"],
-				[" ", ""],
-				["יִ", "!"]
-			],
-			"repair" : [
-				["\\s(\\,|\\)|\\]|\\;)", "$1"],
-				["(\\(\\[)\\s", "$1"],
-				["([^\\.])\\s(\\.)", "$1$2"]
-			]
-		}
-	},
-	{
-		"name" : "en",
-		"config" : {
-			"sentencejoin_model" : "/var/www/html/experiment/sentence-join/toy-model",
-			"join_words" : [
-			],
-			"absolute_eof" : [
-			],
-			"normalize" : [
-			],
-			"repair" : [
-			]
-		}
-	}
-	
-]
-}
