--- Test/html-out/sample.html
@@ -1,41 +1,0 @@
-<html>
-<head>
-<defaultLang abbr="en" />
-<languages>
-<language abbr="en" percent="75.0" />
-<language abbr="ja" percent="25.0" />
-</languages>
-</head>
-<body>
-<div id="page1" class="page">
-<p id="page1p1" lang="en" fontname="Helvetica">
-A Simple PDF File
-</p>
-<p id="page1p2" lang="en" fontname="Helvetica">
-This is a small demonstration.pdf file -
-</p>
-<p id="page1p3" lang="en" fontname="Helvetica">
-just for use in the Virtual Mechanics tutorials. More text. And more text. And more text. And more text. And more text.
-</p>
-<p id="page1p4" lang="ja" fontname="Helvetica">
-And more text. And more text. And more text. And more text. And more text. And more text. Boring, zzzzz. And more text. And more text. And more text. And more text. And more text. And more text. And more text.
-</p>
-<p id="page1p5" lang="en" fontname="Helvetica">
-And more text. And more text.
-</p>
-<p id="page1p6" lang="en" fontname="Helvetica">
-And more text. And more text. And more text. And more text. And more text. And more text. And more text. Even more. Continued on page 2...
-</p>
-</div>
-<div id="page2" class="page">
-<p id="page2p1" lang="en" fontname="Helvetica">
-Simple PDF File 2
-</p>
-<p id="page2p2" lang="en" fontname="Helvetica">
-...continued from page 1. Yet more text. And more text. And more text.
-And more text. And more text. And more text. And more text. And more text. Oh, how boring typing this stuff. But not as boring as watching paint dry. And more text. And more text. And more text. And more text.
-Boring. More, a little more text. The end, and just as well.
-</p>
-</div>
-</body>
-</html>
