--- Test/html-out/sample2.html
@@ -1,73 +1,0 @@
-<html>
-<head>
-<defaultLang abbr="en" />
-<languages>
-<language abbr="en" percent="100.0" />
-</languages>
-</head>
-<body>
-<div id="page1" class="page">
-<p id="page1p1" lang="en" fontname="BAAAAA+Carlito">
-1 Header sample
-</p>
-<p id="page1p2" lang="en" fontname="CAAAAA+DejaVuSans-Bold">
-Airbus A300: Plane that launched an empire
-</p>
-<p id="page1p3" lang="en" fontname="BAAAAA+Carlito">
-(CNN) — This year, around 4 billion plane passengers will take to the sky -- and most of them will travel in an Airbus or a Boeing.
-These two Goliaths have dominated aerospace manufacturing for half a century, with few competitors coming close.
-While US company Boeing, the world leader, has been around since 1916, its European rival
-Airbus emerged much later.
-In the years after World War II, American companies dominated the aviation industry and so, in the late 1960s, a consortium of European aerospace manufacturers, then known as Airbus
-Industrie, came together to develop a challenger.
-On May 29, 1969, at the Paris Air Show, an historic agreement was signed between the French Minister of Transport Jean Chamant and the German Minister of Economic Affairs Karl
-Schiller.
-Airbus Industrie would develop the world's first wide-body twin-engine jet airliner, the Airbus
-A300. It was a plane that would launch an empire.
-</p>
-<p id="page1p4" lang="en" fontname="CAAAAA+DejaVuSans-Bold">
-Light and efficient
-</p>
-<p id="page1p5" lang="en" fontname="BAAAAA+Carlito">
-There were a couple of innovations that made this plane special. The wide body and raised cabin floor meant it could carry passengers and cargo at the same time, allowing airlines to increase profitability.
-The wings, designed by Hawker Siddeley, gave the airplane enough lift to climb faster and attain a level cruise altitude sooner than other passenger aircraft of the time, giving cabin crew more time for the in-flight service.
-</p>
-<p id="page1p6" lang="en" fontname="BAAAAA+Carlito">
-It could hold 270 passengers and, with a range of 1,200 nautical miles, was designed to serve the short- to medium-haul market.
-Because it had two engines rather than the standard three, it was lighter and more efficient than its contemporary US rivals, such as the Lockheed L-1011 TriStar and the McDonnell
-Douglas DC-10.
-</p>
-<p id="page1p7" lang="en" fontname="CAAAAA+DejaVuSans-Bold">
-Construction
-</p>
-<p id="page1p8" lang="en" fontname="BAAAAA+Carlito">
-Construction was a pan-European exercise.
-The wings were constructed in the UK, then transported to Bremen, Germany. to be fitted with moving surfaces. They were then shipped to nearby Lemwerder, Germany, where they -- along with the rear fuselage sections -- were flown to the final assembly line in Toulouse, France.
-The engines came from the US, the tail assemblies from Spain, while the central wingbox and nose were built in the French towns of St Nazaire and Nantes.
-The road convoy of parts to the assembly line attracted curious crowds -- an Airbus tradition that would continue through the decades, with people regularly turning out to see the components arrive for the mammoth A380.
-</p>
-<p id="page1p9" lang="en" fontname="CAAAAA+DejaVuSans-Bold">
-Unveiling
-</p>
-<p id="page1p10" lang="en" fontname="BAAAAA+Carlito">
-The A300B was unveiled to the French public on September 28, 1972, in a joint ceremony with another European plane which was to make history: the legendary, supersonic Concorde.
-The A300B made its first flight one month later, in October 1972, landing in high crosswinds that tested the mettle of pilots Max Fischl and Bernard Ziegler.
-</p>
-</div>
-<div id="page2" class="page">
-<p id="page2p1" lang="en" fontname="BAAAAA+Carlito">
-2 Header sample
-</p>
-<p id="page2p2" lang="en" fontname="BAAAAA+Carlito">
-The plane entered service with Air France in May 1974. Sales were slow at the start, and it was unclear whether the European start-up would survive.
-</p>
-<p id="page2p3" lang="en" fontname="BAAAAA+Carlito">
-Three-engine planes were still the industry standard, and there was skepticism whether two engines could safely carry passengers over long distances.
-</p>
-<p id="page2p4" lang="en" fontname="BAAAAA+Carlito">
-A lifeline appeared in 1977 when the US-based
-Eastern Airlines gave them a shot by leasing four planes for six months. The company ended up buying those craft and ordering 19 more.
-</p>
-</div>
-</body>
-</html>
